pyproject.toml
[project]
name = "pytritech"
description = "A Python interface to Tritech Gemini Sonar files."
version = "1.0.0"
authors = [
    { name = "Benjamin James Blundell", email = "me@benjamin.computer" }
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]

dependencies = [
    "pytz",
    "isal",
    "zipfile-isal"
]

[project.optional-dependencies]
dev = [
    "tox",
    "pre-commit",
    "bump2version",
    "mkdocs",
    "mkdocstrings-python",
    "pyinstrument",
    "GitPython",
    "pillow",
]

test = [
    "pytest"
]

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.2,<4"]

[project.urls]
"Homepage" = "https://github.com/onidaito/pytritech"
"Bug Tracker" = "https://github.com/onidaito/pytritech/issues"